refactor(test): shared suites with native runners, fix serve.ts Response guard
Replace 10 duplicated per-category test files (Node + Deno variants
differing only in import line) with 7 shared suite modules in
tests/suites/ and 2 thin platform runners in tests/runners/.

Suites: lifecycle, errors, stress, events, sessions, keys, discovery.
Each exports a function taking { createNode, test, assert, ... } — the
runner binds platform-native primitives (node:test / Deno.test).

Coverage additions from existing e2e/smoke suites:
- Key sign/verify round-trip, PublicKey.fromString, SecretKey.fromBytes
- URL scheme validation (https://, http:// rejection)
- Peer-Id header format (base32, consistent)
- No-stderr on valid responses
- browse(), pathChanges(), advertise() API shape

fix(shared): guard against non-Response handler return in serve.ts.
The handler callback is user code and may return a non-Response value.
Previously this caused 'res.headers is not iterable' + a cascading
finishBody error. Now throws TypeError before accessing Response
properties, caught by the existing onError fallback.

63 tests pass on both Node and Deno.
