--- tests/stress/test-deno.ts
@@ -1,259 +1,0 @@
-/**
- * iroh-http stress tests — Deno
- *
- * Tests high concurrency, large payloads, rapid lifecycle churn,
- * and resource behavior under load.
- *
- * Usage:
- *   deno run -A tests/stress/test-deno.ts
- */
-
-import { createNode } from "../../packages/iroh-http-deno/mod.ts";
-import {
-  suite, test, assert, assertEqual, run,
-} from "../harness.mjs";
-
-suite("stress");
-
-// ── Concurrent fetch ────────────────────────────────────────────────────────
-
-test("50 concurrent GETs all succeed", async () => {
-  const server = await createNode();
-  const client = await createNode();
-  const { id: serverId, addrs: serverAddrs } = await server.addr();
-
-  server.serve({}, () => new Response("ok"));
-
-  const results = await Promise.all(
-    Array.from({ length: 50 }, (_, i) =>
-      client.fetch(serverId, `/concurrent-${i}`, { directAddrs: serverAddrs }).then(async (res) => {
-        const body = await res.text();
-        return { status: res.status, body };
-      })
-    )
-  );
-
-  for (let i = 0; i < results.length; i++) {
-    assertEqual(results[i].status, 200, `request ${i} status`);
-    assertEqual(results[i].body, "ok", `request ${i} body`);
-  }
-
-  await server.close();
-  await client.close();
-});
-
-test("20 concurrent POSTs with body all echo correctly", async () => {
-  const server = await createNode();
-  const client = await createNode();
-  const { id: serverId, addrs: serverAddrs } = await server.addr();
-
-  server.serve({}, async (req) => {
-    const body = await req.text();
-    return new Response(body);
-  });
-
-  const results = await Promise.all(
-    Array.from({ length: 20 }, (_, i) =>
-      client.fetch(serverId, "/echo", {
-        method: "POST",
-        body: `message-${i}`,
-        directAddrs: serverAddrs,
-      }).then(async (res) => {
-        const body = await res.text();
-        return { status: res.status, body, index: i };
-      })
-    )
-  );
-
-  for (const r of results) {
-    assertEqual(r.status, 200, `POST ${r.index} status`);
-    assertEqual(r.body, `message-${r.index}`, `POST ${r.index} body`);
-  }
-
-  await server.close();
-  await client.close();
-});
-
-// ── Sequential fetch (connection reuse) ─────────────────────────────────────
-
-test("100 sequential GETs all succeed", async () => {
-  const server = await createNode();
-  const client = await createNode();
-  const { id: serverId, addrs: serverAddrs } = await server.addr();
-
-  let count = 0;
-  server.serve({}, () => {
-    count++;
-    return new Response(`req-${count}`);
-  });
-
-  for (let i = 0; i < 100; i++) {
-    const res = await client.fetch(serverId, `/seq-${i}`, { directAddrs: serverAddrs });
-    assertEqual(res.status, 200, `seq ${i} status`);
-    const body = await res.text();
-    assert(body.startsWith("req-"), `seq ${i} body prefix`);
-  }
-
-  assertEqual(count, 100, "handler call count");
-
-  await server.close();
-  await client.close();
-});
-
-// ── Large body transfers ────────────────────────────────────────────────────
-
-test("1MB body round-trip", async () => {
-  const server = await createNode();
-  const client = await createNode();
-  const { id: serverId, addrs: serverAddrs } = await server.addr();
-
-  server.serve({}, async (req) => {
-    const buf = await req.arrayBuffer();
-    return new Response(new Uint8Array(buf));
-  });
-
-  const MB = 1024 * 1024;
-  const payload = new Uint8Array(MB);
-  // Fill with pattern for verification
-  for (let i = 0; i < MB; i++) payload[i] = i & 0xff;
-
-  const res = await client.fetch(serverId, "/echo", {
-    method: "POST",
-    body: payload,
-    directAddrs: serverAddrs,
-  });
-
-  assertEqual(res.status, 200, "status");
-  const buf = await res.arrayBuffer();
-  assertEqual(buf.byteLength, MB, "response size");
-
-  // Verify first and last bytes
-  const view = new Uint8Array(buf);
-  assertEqual(view[0], 0, "first byte");
-  assertEqual(view[1], 1, "second byte");
-  assertEqual(view[MB - 1], (MB - 1) & 0xff, "last byte");
-
-  await server.close();
-  await client.close();
-});
-
-test("5 concurrent 256KB transfers", async () => {
-  const server = await createNode();
-  const client = await createNode();
-  const { id: serverId, addrs: serverAddrs } = await server.addr();
-
-  server.serve({}, async (req) => {
-    const buf = await req.arrayBuffer();
-    return new Response(String(buf.byteLength));
-  });
-
-  const SIZE = 256 * 1024;
-  const results = await Promise.all(
-    Array.from({ length: 5 }, () =>
-      client.fetch(serverId, "/echo-length", {
-        method: "POST",
-        body: new Uint8Array(SIZE),
-        directAddrs: serverAddrs,
-      }).then(async (res) => {
-        const body = await res.text();
-        return { status: res.status, bodyLength: parseInt(body, 10) };
-      })
-    )
-  );
-
-  for (let i = 0; i < results.length; i++) {
-    assertEqual(results[i].status, 200, `transfer ${i} status`);
-    assertEqual(results[i].bodyLength, SIZE, `transfer ${i} size`);
-  }
-
-  await server.close();
-  await client.close();
-});
-
-// ── Rapid lifecycle churn ───────────────────────────────────────────────────
-
-test("20 rapid create/close cycles with fetch", async () => {
-  for (let i = 0; i < 20; i++) {
-    const server = await createNode();
-    const client = await createNode();
-    const { id: serverId, addrs: serverAddrs } = await server.addr();
-
-    server.serve({}, () => new Response(`cycle-${i}`));
-
-    const res = await client.fetch(serverId, "/", { directAddrs: serverAddrs });
-    assertEqual(res.status, 200, `cycle ${i} status`);
-    const body = await res.text();
-    assertEqual(body, `cycle-${i}`, `cycle ${i} body`);
-
-    await server.close();
-    await client.close();
-  }
-});
-
-// ── Multiple concurrent nodes ───────────────────────────────────────────────
-
-test("5 independent node pairs concurrently", async () => {
-  const pairs = await Promise.all(
-    Array.from({ length: 5 }, async (_, i) => {
-      const server = await createNode();
-      const client = await createNode();
-      const { id: serverId, addrs: serverAddrs } = await server.addr();
-      server.serve({}, () => new Response(`pair-${i}`));
-      return { server, client, serverId, serverAddrs, index: i };
-    })
-  );
-
-  // All pairs fetch concurrently
-  const results = await Promise.all(
-    pairs.map(async ({ client, serverId, serverAddrs, index }) => {
-      const res = await client.fetch(serverId, "/", { directAddrs: serverAddrs });
-      const body = await res.text();
-      return { index, status: res.status, body };
-    })
-  );
-
-  for (const r of results) {
-    assertEqual(r.status, 200, `pair ${r.index} status`);
-    assertEqual(r.body, `pair-${r.index}`, `pair ${r.index} body`);
-  }
-
-  // Cleanup all
-  await Promise.all(
-    pairs.flatMap(({ server, client }) => [server.close(), client.close()])
-  );
-});
-
-// ── Mixed methods under load ────────────────────────────────────────────────
-
-test("30 concurrent mixed-method requests", async () => {
-  const server = await createNode();
-  const client = await createNode();
-  const { id: serverId, addrs: serverAddrs } = await server.addr();
-
-  server.serve({}, (req) => {
-    return new Response(req.method);
-  });
-
-  const methods = ["GET", "POST", "PUT", "PATCH", "DELETE", "OPTIONS"];
-  const results = await Promise.all(
-    Array.from({ length: 30 }, (_, i) => {
-      const method = methods[i % methods.length];
-      return client.fetch(serverId, `/method-${i}`, { method, directAddrs: serverAddrs }).then(async (res) => {
-        const body = await res.text();
-        return { method, body, status: res.status };
-      });
-    })
-  );
-
-  for (const r of results) {
-    assertEqual(r.status, 200, `${r.method} status`);
-    assertEqual(r.body, r.method, `${r.method} echo`);
-  }
-
-  await server.close();
-  await client.close();
-});
-
-// ── Run ─────────────────────────────────────────────────────────────────────
-const code = await run();
-Deno.exit(code);
